move user story 36 from edit spec to show spec, delete edit spec
